make browser-safe imports
