+++ packages/auth/index.browser.ts
@@ -1,0 +1,1 @@
+export { checkPermissions } from './next-auth/session.browser'
