+++ packages/auth/next-auth/session.server.ts
@@ -1,9 +1,8 @@
 import { type GetServerSidePropsContext, type NextApiRequest, type NextApiResponse } from 'next'
-import { getServerSession as nextauthServerSession, type Session } from 'next-auth'
-
-import { type Permission } from '@weareinreach/db/generated/permission'
+import { getServerSession as nextauthServerSession } from 'next-auth'
 
 import { authOptions } from './auth-options'
+import { checkPermissions, type CheckPermissionsParams } from './session.browser'
 
 interface SSRContext {
 	req: GetServerSidePropsContext['req']
@@ -17,20 +16,6 @@ type ServerContext = SSRContext | ApiContext
 
 export const getServerSession = async (ctx: ServerContext) => {
 	return nextauthServerSession(ctx.req, ctx.res, authOptions)
-}
-interface CheckPermissionsParams {
-	session: Session | null
-	permissions: Permission | Permission[]
-	has: 'all' | 'some'
-}
-export const checkPermissions = ({ session, permissions, has = 'all' }: CheckPermissionsParams) => {
-	if (!session) return false
-	if (session.user.permissions.includes('root')) return true
-	const userPerms = new Set(session.user.permissions)
-	const permsToCheck = Array.isArray(permissions) ? permissions : [permissions]
-	return has === 'all'
-		? permsToCheck.every((perm) => userPerms.has(perm))
-		: permsToCheck.some((perm) => userPerms.has(perm))
 }
 
 interface CheckServerPermissions extends Omit<CheckPermissionsParams, 'session'> {
